✨[feat]: MainCollectionView 타입 생성 및 로직 분리
